requirements/common.txt
Flask==1.1.*
Flask-GraphQL==2.0.*
Flask-Migrate==2.5.*
Flask-SQLAlchemy==2.4.*
graphene==2.1.*
graphene-sqlalchemy==2.2.*
gunicorn==20.0.*
psycopg2==2.8.*
python-dotenv==0.13.*
